GUACAMOLE-394: Separate definition of records and record sets into generalized interface.
