+++ guacamole-ext/src/main/java/org/apache/guacamole/net/auth/ConnectionRecord.java
@@ -19,13 +19,11 @@
 
 package org.apache.guacamole.net.auth;
 
-import java.util.Date;
-
 /**
  * A logging record describing when a user started and ended usage of a
  * particular connection.
  */
-public interface ConnectionRecord {
+public interface ConnectionRecord extends ActivityRecord {
 
     /**
      * Returns the identifier of the connection associated with this
@@ -71,49 +69,5 @@ public interface ConnectionRecord {
      *     was accessed directly.
      */
     public String getSharingProfileName();
-
-    /**
-     * Returns the date and time the connection began.
-     *
-     * @return The date and time the connection began.
-     */
-    public Date getStartDate();
-
-    /**
-     * Returns the date and time the connection ended, if applicable.
-     *
-     * @return The date and time the connection ended, or null if the
-     *         connection is still running or if the end time is unknown.
-     */
-    public Date getEndDate();
-
-    /**
-     * Returns the hostname or IP address of the remote host that used the
-     * connection associated with this record, if known. If the hostname or IP
-     * address is not known, null is returned.
-     *
-     * @return
-     *     The hostname or IP address of the remote host, or null if this
-     *     information is not available.
-     */
-    public String getRemoteHost();
-
-    /**
-     * Returns the name of the user who used or is using the connection at the
-     * times given by this connection record.
-     *
-     * @return The name of the user who used or is using the associated
-     *         connection.
-     */
-    public String getUsername();
-
-    /**
-     * Returns whether the connection associated with this record is still
-     * active.
-     *
-     * @return true if the connection associated with this record is still
-     *         active, false otherwise.
-     */
-    public boolean isActive();
 
 }
